solutions/5_solution/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mc-solution4"
version = "0.1.0"
description = "Solution 5 TigerBeetle + Restate showcase for metered-compute API"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi==0.129.0",
  "uvicorn[standard]==0.40.0",
  "redis==6.4.0",
  "asyncpg==0.31.0",
  "pydantic==2.12.5",
  "pydantic-settings==2.13.0",
  "structlog==25.5.0",
  "prometheus-client==0.24.1",
  "uuid6==2025.0.1",
  "tigerbeetle==0.16.68",
  "restate-sdk==0.14.2",
  "httpx==0.28.1",
  "orjson==3.11.7",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.1",
  "pytest-asyncio==1.0.0",
  "pytest-cov==7.0.0",
  "httpx==0.28.1",
  "ruff==0.11.12",
  "mypy==1.16.0",
  "bandit==1.8.6",
  "pip-audit==2.9.0",
  "detect-secrets==1.5.0",
  "radon==6.0.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 120
extend-exclude = ["worklog"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
pretty = true
show_error_codes = true
files = ["src", "tests"]

[[tool.mypy.overrides]]
module = ["asyncpg", "asyncpg.*", "structlog", "restate", "restate.*", "tigerbeetle", "tigerbeetle.*"]
ignore_missing_imports = true

# Restate SDK decorator (@task_service.handler) is untyped; ctx.run() returns Any.
[[tool.mypy.overrides]]
module = ["solution5.workflows"]
disallow_untyped_decorators = false

# redis.asyncio stubs type from_url as untyped.
[[tool.mypy.overrides]]
module = ["solution5.app", "solution5.cache"]
disallow_untyped_calls = false

[tool.coverage.run]
branch = true
source = ["src/solution5"]

[tool.coverage.report]
skip_covered = false
show_missing = true
precision = 1

[tool.bandit]
exclude_dirs = ["tests", "worklog"]
skips = ["B104"]
